// src/API/APIEntryPoints.ts
export namespace Session {
    /**
     * Create Request object to login into nicovideo.jp.
     * @param {string} email An E-mail used to login.
     * @param {string} password A password used to login.
     * @returns {Axios.AxiosRequestConfig}
     */
    export function createLoginRequest(email: string, password: string): Axios.AxiosRequestConfig {
        return {
            method: "POST",
            url: APIUrl.LOGIN,
            data: {
                mail_tel: email,
                password: password,
            },
            withCredentials: true,
        };
    }

    /**
     * Create Request Object for Multi-Factor Authentication
     * @param {string} url The URL to post
     * @param {string} oneTimePad Required One-Time-Pad generated by application or sent by e-mail
     * @param {boolean} trust Keep device trusted or not
     * @param {string} deviceName
     * @returns {AxiosRequestConfig}
     */
    export function createMultiFactorAuthenticationRequest(url: string, oneTimePad: string, trust: boolean, deviceName?: string): Axios.AxiosRequestConfig {
        return {
            method: "POST",
            url: url,
            data: {
                otp: oneTimePad,
                is_mfa_trusted_device: trust,
                device_name: deviceName,
            },
            withCredentials: true,
        };
    }

    /**
     * Create Request object to logout from nicovideo.jp.
     * @returns {Axios.AxiosRequestConfig}
     * @todo check is collect.
     */
    export function createLogoutRequest(): Axios.AxiosRequestConfig {
        return {
            method: "GET",
            url: APIUrl.LOGOUT,
            withCredentials: true,
        };
    }

    export function createCheckActiveRequest(): Axios.AxiosRequestConfig {
        return {
            method: "GET",
            url: APIUrl.MYLIST_DEFAULT,
            withCredentials: true,
        };
    }
}